backend/requirements.txt
fastapi>=0.116.0
python-multipart>=0.0.12
uvicorn>=0.35.0
torch>=2.8.0
Pillow>=11.3.0
numpy>=2.3.0

